requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asyncio==3.4.3
azure-core==1.34.0
azure-identity==1.23.0
azure-kusto-data==5.0.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
ijson==3.4.0
mcp==1.9.4
msal==1.32.3
msal-extensions==1.3.1
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
requests==2.32.4
six==1.17.0
sniffio==1.3.1
sse-starlette==2.3.6
starlette==0.47.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
python-dotenv
